feat(config): 增强配置管理功能
- 在配置初始化中添加数据正确性检查，确保系统端口和模式符合要求。
- 修改系统配置中的端口类型为 uint16，以支持更大的端口范围。
- 更新路由启动方法，允许通过参数指定端口，默认端口设置为 8888。
- 优化配置文件读取逻辑，确保在文件缺失时使用默认配置。

+++ internal/model/other/config.go
@@ -17,6 +17,6 @@ type Database struct {
 
 // System 系统配置
 type System struct {
-	Port int `yaml:"port"`
+	Port uint16 `yaml:"port"`
 	Mode string `yaml:"mode"`
 }
